+++ quantify_scheduler/backends/qblox/constants.py
@@ -56,6 +56,8 @@ NUMBER_OF_REGISTERS: int = 64
 """Number of registers available in the Qblox sequencers."""
 MAX_SAMPLE_SIZE_ACQUISITIONS: int = 16384
 """Maximal amount of acquisition datapoints returned."""
+MAX_SAMPLE_SIZE_WAVEFORMS: int = 16384
+"""Maximal amount of samples in the waveforms to be uploaded to a sequencer."""
 MAX_NUMBER_OF_BINS: int = 131072
 """Number of bins available in the Qblox sequencers."""
 GENERIC_IC_COMPONENT_NAME: str = "generic"